moved msva payload generator
